+++ EventFilter/Utilities/BuildFile.xml
@@ -1,6 +1,10 @@
 <use   name="FWCore/MessageLogger"/>
 <use   name="FWCore/ServiceRegistry"/>
+<use   name="FWCore/Framework"/>
+<use   name="FWCore/Sources"/>
 <use   name="EventFilter/FEDInterface"/>
+<use   name="DataFormats/FEDRawData"/>
+<use   name="IOPool/Streamer"/>
 <use   name="curl"/>
 <export>
   <lib   name="1"/>
