+++ EventFilter/Utilities/plugins/BuildFile.xml
@@ -5,8 +5,8 @@
   <use   name="FWCore/Sources"/>
   <use   name="FWCore/Utilities"/>
   <use   name="IOPool/Streamer"/>
-  <use   name="EventFilter/Utilities"/>
   <use   name="DataFormats/FEDRawData"/>
+  <use   name="EventFilter/Utilities"/>
   <use   name="root"/>
   <use   name="boost"/>
   <use   name="CLHEP"/>
